Rename trainer_task.py -> task.py; establish task.py convention
Align with internal SDK convention where every workflow task module
uses task.py as the primary file. The __init__.py public API is
unchanged — callers are unaffected.

- trainer_task.py -> task.py
- tests/trainer_task_test.py -> tests/task_test.py
- __init__.py import updated accordingly
- patch target string in task_test.py updated
- _dataset.py docstring reference updated

Requested by @kenns29 in PR review; tracked in #1428 for pusher.
